requirements.txt
PyQt5==5.10.1
PyYAML==3.13